+++ org.eclipse.jdt.core/compiler/org/eclipse/jdt/internal/compiler/impl/ByteConstant.java
@@ -16,25 +16,25 @@ public ByteConstant(byte value) {
 	this.value = value;
 }
 public byte byteValue() {
-	return (byte) value;
+	return value;
 }
 public char charValue() {
 	return (char) value;
 }
 public double doubleValue() {
-	return (double) value;
+	return value;
 }
 public float floatValue() {
-	return (float) value;
+	return value;
 }
 public int intValue() {
-	return (int) value;
+	return value;
 }
 public long longValue() {
-	return (long) value;
+	return value;
 }
 public short shortValue() {
-	return (short) value;
+	return value;
 }
 public String stringValue() {
 	//spec 15.17.11
