+++ org.eclipse.jdt.core/compiler/org/eclipse/jdt/internal/compiler/impl/IntConstant.java
@@ -27,19 +27,19 @@ public class IntConstant extends Constant {
 	}
 	
 	public double doubleValue() {
-		return (double) value;
+		return value;
 	}
 	
 	public float floatValue() {
-		return (float) value;
+		return value;
 	}
 	
 	public int intValue() {
-		return (int) value;
+		return value;
 	}
 	
 	public long longValue() {
-		return (long) value;
+		return value;
 	}
 	
 	public short shortValue() {
